chore: ui optimizations & repackage

+++ app/src/main/java/com/example/roberto/recyclerviewsample/utils/PixelConverters.kt
@@ -2,7 +2,7 @@ package com.example.roberto.recyclerviewsample.utils
 
 import android.content.res.Resources
 
-val Int.dp: Int
+val Float.dp: Int
     get() = (this / Resources.getSystem().displayMetrics.density).toInt()
-val Int.px: Int
+val Float.px: Int
     get() = (this * Resources.getSystem().displayMetrics.density).toInt()
